requirements.txt
Pillow
Tkinter
Requests
